refactor: change plural names
